Move doc-link helpers from interfaces to cli/doc_links.py
Keep interfaces/help_topic.py as pure plugin-facing models (DocFile just
declares the source_url field). The repo-identity + release-version URL builder
(imbue_mngr_doc_url) and the relative/anchor link rewriter now live together in
the new cli/doc_links.py -- application/CLI knowledge that doesn't belong in the
interfaces boundary. Built-in topics, the mngr_usage plugin, and the help
renderer import from there.
