+++ libs/mngr/imbue/mngr/interfaces/help_topic.py
@@ -18,45 +18,12 @@ All metadata (key, description, aliases, see-also) is declared explicitly --
 nothing is inferred by parsing the body.
 """
 
-from importlib.metadata import PackageNotFoundError
-from importlib.metadata import version
 from pathlib import Path
 from typing import assert_never
 
 from pydantic import Field
 
 from imbue.imbue_common.frozen_model import FrozenModel
-
-_IMBUE_MNGR_REPO_URL = "https://github.com/imbue-ai/mngr"
-
-
-def _imbue_mngr_release_ref() -> str:
-    """The git ref to pin doc links to: the installed mngr release's tag, else ``main``.
-
-    Released wheels ship their docs in lockstep with their version tag (e.g.
-    ``v0.2.9``), so links pinned to that tag resolve to exactly the docs the user
-    has installed. Falls back to ``main`` when the distribution version can't be
-    read (e.g. mngr isn't installed as a package). Caveat: in a source checkout
-    whose version predates a not-yet-released doc, that doc's link can 404 until
-    the next release -- inherent to version-pinning, and harmless for real
-    installs (where version and shipped docs always match).
-    """
-    try:
-        return f"v{version('imbue-mngr')}"
-    except PackageNotFoundError:
-        return "main"
-
-
-def imbue_mngr_doc_url(repo_relative_path: str) -> str:
-    """GitHub blob URL for a doc shipped in the imbue-ai/mngr repo, pinned to the installed release.
-
-    ``repo_relative_path`` is the doc's path from the repo root (e.g.
-    ``"libs/mngr/docs/concepts/idle_detection.md"``). In-repo topic providers
-    (mngr's built-ins and the mngr_usage plugin) use this to populate
-    :attr:`DocFile.source_url`, so relative links in those docs render as working
-    GitHub URLs when shown in an interactive terminal.
-    """
-    return f"{_IMBUE_MNGR_REPO_URL}/blob/{_imbue_mngr_release_ref()}/{repo_relative_path}"
 
 
 class InlineContent(FrozenModel):
