+++ libs/mngr_usage/imbue/mngr_usage/plugin.py
@@ -15,10 +15,10 @@ from pathlib import Path
 import click
 
 from imbue.mngr import hookimpl
+from imbue.mngr.cli.doc_links import imbue_mngr_doc_url
 from imbue.mngr.config.plugin_registry import register_plugin_config
 from imbue.mngr.interfaces.help_topic import DocFile
 from imbue.mngr.interfaces.help_topic import TopicHelpPage
-from imbue.mngr.interfaces.help_topic import imbue_mngr_doc_url
 from imbue.mngr_usage.cli import usage
 from imbue.mngr_usage.data_types import UsagePluginConfig
